feat: 우측 트렌드 메뉴 구현
